Move component logic
